sourcecode/server/requirements.txt
flask==2.3.3
flask-cors==4.0.0
boto3==1.28.40
python-dotenv==1.0.0
werkzeug==2.3.7
reportlab==4.0.4
PyPDF2==3.0.1
uuid==1.30
pillow==10.0.0
weasyprint==59.0
requests==2.31.0
